Major refactor
I refactored databasemanager. It was  way to big and unreadable.

I broke it all up into a bunch of services and got rid of it, now its far easier to figure out whats going on.
