+++ Assets/scripts/DataBase/DatabaseConnectionManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 53eb44d34028caf45bac5323d454b2fe
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
