+++ Assets/scripts/DataBase/Services.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: aa341d0fcb3861e458c525d88cb522ab
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
