+++ Assets/scripts/DataBase/Services/CategoryService.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 70ce4b056d20034468ff6281388f3a4d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
