+++ Assets/scripts/DataBase/Services/InventoryService.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bd5f35d659a2ef44dac36627fba67573
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
