+++ Assets/scripts/DataBase/Services/ProductCategoryQueryService.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 70b4d3bda031f9c4fa79e0b92aadceae
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
