+++ Assets/scripts/DataBase/Services/ProductCategoryService.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 07ad58f97836dec4780c98c962cbfdba
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
