+++ Assets/scripts/DataBase/Services/ProductService.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: da470b7964b7bb14eb238b0c5a2ad8c2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
