+++ Assets/scripts/DataBase/Services/ShelfService.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 63a0b5a2a665baf45b9ef20a2728ff72
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
