+++ Assets/scripts/DataBase/Services/ShelvingUiService.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f56f7653ec3868744970e85bf9df82f9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
