+++ Assets/scripts/DataBase/Services/TableManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: aa15ad3ad4ef7314dbed25fb0c4f507b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
